Move out extension logic to external function
