+++ packages/shared-src/src/models/fhir/Patient.js
@@ -14,15 +14,13 @@ import {
 } from '../../constants';
 import {
   FhirAddress,
-  FhirCodeableConcept,
-  FhirCoding,
   FhirContactPoint,
-  FhirExtension,
   FhirHumanName,
   FhirIdentifier,
   FhirPatientLink,
   FhirReference,
 } from '../../services/fhirTypes';
+import { nzEthnicity } from './extensions';
 
 export class FhirPatient extends FhirResource {
   static init(options, models) {
@@ -172,39 +170,6 @@ export class FhirPatient extends FhirResource {
 
 function compactBy(array, access = identity) {
   return array.filter(access);
-}
-
-function nzEthnicity(patient) {
-  function getEthnicity(ethnicityId) {
-    switch (ethnicityId) {
-      case 'ethnicity-Fiji':
-        return { code: '36111', display: 'Fijian/iTaukei' };
-      case 'ethnicity-FID':
-        return { code: '43112', display: 'Fijian Indian' };
-      case null:
-        return { code: '99999', display: 'Not Stated' };
-      default:
-        return { code: '61199', display: 'Other Ethnicity nec' };
-    }
-  }
-
-  if (!config.localisation.data.features.fhirNewZealandEthnicity) return [];
-  const { code, display } = getEthnicity(patient?.additionalData?.ethnicityId);
-
-  return [
-    new FhirExtension({
-      url: 'http://hl7.org.nz/fhir/StructureDefinition/nz-ethnicity',
-      valueCodeableConcept: new FhirCodeableConcept({
-        coding: [
-          new FhirCoding({
-            system: config.hl7.dataDictionaries.ethnicityId,
-            code,
-            display,
-          }),
-        ],
-      }),
-    }),
-  ];
 }
 
 function extension(patient) {
